+++ src/my_first_adventure_game/game/scenes/__init__.py
@@ -2,6 +2,7 @@ from my_first_adventure_game.game.scenes.defeat_scene import DefeatScene
 from my_first_adventure_game.game.scenes.dialogue_scene import DialogueScene
 from my_first_adventure_game.game.scenes.gameplay_scene import GameplayScene
 from my_first_adventure_game.game.scenes.pause_scene import PauseScene
+from my_first_adventure_game.game.scenes.profile_scene import ProfileScene
 from my_first_adventure_game.game.scenes.title_scene import TitleScene
 from my_first_adventure_game.game.scenes.victory_scene import VictoryScene
 
@@ -10,6 +11,7 @@ __all__ = [
     "DialogueScene",
     "GameplayScene",
     "PauseScene",
+    "ProfileScene",
     "TitleScene",
     "VictoryScene",
 ]